Added giss_data.py module.
This is work-in-progress, but the basics are, I think, fairly solid.
This module is not used by any of the current checked in code, but
I felt I should make this available to others for comment, etc.

I have code that works with this and, I think, is more readable as a result.


git-svn-id: http://ccc-gistemp.googlecode.com/svn/trunk@230 a3c903a6-8ee9-11dd-8a4e-8714b3728bc5