Modified step0 to use 12-digit station IDs throughout, to use float data throughout (so there is no rounding at all, until producing the output file), with a BAD value rather than None (for consistency with other steps), and to produce a (dict, series) iterator (similar to the ones through steps 1 and 2).  Also separated the main algorithm (step0()) from the code to produce the output file (write_data()).  write_data() may be subsequently moved out to a different file to isolate data processing in this file only.  Also rearranged the file so that reading input datasets and metadata are in early sections, separate from the (short) data processing section (which mainly exists to adjust USHCN data to match GHCN data means).
The change to step0 makes a tiny (and statistically irrelevant) difference to the end results, in monthly data for boxes covering North America, because the adjustment of USHCN data to match GHCN means now happens in floating-point.

Modified step1 to separate out the code which reads the step0 output file (see above).  This will make it trivial to wire together step0.step0() and step1.step1(), and easy to move I/O code into a separate module in due course.


git-svn-id: http://ccc-gistemp.googlecode.com/svn/trunk@251 a3c903a6-8ee9-11dd-8a4e-8714b3728bc5